[#170198513] Move listing_service to soql folder, hide listings from pre-lottery
- Renamed "listings" to pre_lottery_listings to be clearer about what listings we're querying for.
- VCR cassettes had to be updated because moving the functions to the new file increased the number of spaces in the query, causing the cassettes to no longer match
