+++ app/controllers/listings/lease_ups/applications_controller.rb
@@ -6,7 +6,7 @@ module Listings::LeaseUps
     before_action :authenticate_user!
 
     def index
-      full_listing = listing_service.listing(params[:lease_up_id])
+      full_listing = soql_listing_service.listing(params[:lease_up_id])
 
       raise Force::RecordNotFound, "Listing #{params[:lease_up_id]} not found" unless full_listing.present?
       @listing = compacted_listing(full_listing)
@@ -16,8 +16,8 @@ module Listings::LeaseUps
       Force::LeaseUpService.new(current_user)
     end
 
-    def listing_service
-      Force::ListingService.new(current_user)
+    def soql_listing_service
+      Force::Soql::ListingService.new(current_user)
     end
 
     def compacted_listing(full_listing)
